apparently fixede

+++ model/robot.h
@@ -1,7 +1,7 @@
 #ifndef ROBOT_INCLUDED
 #define ROBOT_INCLUDED
 
-float slideE(int m, int n);
+double slideE(int m, int n);
 bool inLimit(int voltage);
 
 #endif
